svn merge -r6645:6689 https://eris.llnl.gov/svn/slurm/branches/slurm-0-6-branch

+++ src/plugins/jobacct/log/jobacct_log.c
@@ -122,7 +122,7 @@ typedef struct _stats_msg {
 	_stats_msg_type_t	msg_type;
 	uint32_t	 	jobid;   /* in network order! */
 	uint32_t		stepid;  /* ditto */
-	uint16_t		datalen; /* ditto */
+	uint32_t		datalen; /* ditto (and make 'data' aligned) */
 	char			data[MAX_MSG_SIZE];
 } _stats_msg_t;
